feat: add fractional GPU support for AWS and improve Vultr region handling
Enable sub-GPU allocation (e.g. A100(count=1/4)) on AWS by detecting
fractional instance families (g6f/gr6f), auto-injecting the NVIDIA GRID
driver plugin, and selecting Ubuntu AMIs instead of DLAMIs. The offer
pipeline now filters by accelerator_count, and the pool actor merges
provider-injected plugins before image transforms.

Vultr region is now optional — when unset, offers from all available
regions are listed and expanded per-region. Also extracts VRAM from
cloud GPU plan IDs for more accurate offer matching.

+++ skyward/providers/aws/types.py
@@ -40,7 +40,7 @@ class InstanceResources:
     vcpus: int
     memory_mb: int
     architecture: str
-    gpu_count: int = 0
+    gpu_count: float = 0
     gpu_model: str = ""
     gpu_vram_mb: int = 0
     network_bandwidth_gbps: float = 0.0
@@ -63,7 +63,7 @@ class InstanceSpec:
     vcpus: int
     memory_gb: float
     architecture: str
-    gpu_count: int = 0
+    gpu_count: float = 0
     gpu_model: str = ""
     gpu_vram_gb: float = 0.0
     network_bandwidth_gbps: float = 0.0
